+++ packages/react/src/index.js
@@ -1,2 +1,3 @@
 export { Button } from './components/Button/Button';
 export { usePageVisibility } from './hooks/usePageVisibility';
+export * from './magic-auth';