+++ packages/react/src/magic-auth/Context.js
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+const DEFAULT_CONTEXT = null;
+
+export const Context = React.createContext(DEFAULT_CONTEXT);
+
+Context.DEFAULT = DEFAULT_CONTEXT;