+++ packages/react/src/magic-auth/index.js
@@ -1,0 +1,2 @@
+export { useMagicAuth } from './useMagicAuth';
+export { MagicAuthProvider } from './MagicAuthProvider';