+++ sites/magic.iamnoah.com/src/components/AuthWatchLoginToken.js
@@ -1,24 +1,28 @@
 import * as React from 'react';
 
-import { useAuth } from 'src/components/AuthProvider';
+import { useMagicAuth } from '@magusn/react';
 import graphql from 'src/client/graphql/queries';
 
 function WatchLoginToken({ loginTokenId, handleLogout }) {
   const { loginToken, ...result } = graphql.watchLoginToken(loginTokenId);
 
-  React.useEffect(async () => {
-    if (!result.loading && !result.error && !loginToken) {
-      // loginToken is missing, logout immediately
-      console.debug('[AuthProvider]', 'no error and loginToken missing', 'logout');
-      await handleLogout();
+  React.useEffect(() => {
+    async function asyncEffect() {
+      if (!result.loading && !result.error && !loginToken) {
+        // loginToken is missing, logout immediately
+        console.debug('[AuthWatchLoginToken]', 'no error and loginToken missing', 'logout');
+        await handleLogout();
+      }
     }
+
+    asyncEffect();
   }, [result.loading, loginToken]);
 
   return null;
 }
 
 export default function AuthWatchLoginToken() {
-  const auth = useAuth();
+  const auth = useMagicAuth();
 
   if (auth.isLoggedIn) {
     return <WatchLoginToken loginTokenId={auth.loginRequestId} handleLogout={auth.actions.logout} />;