+++ sites/magic.iamnoah.com/CHANGELOG.md
@@ -1,9 +1,17 @@
 # magic.iamnoah.com
 
-## 0.4.1
+## 0.4.2
+
 ### Patch Changes
 
+- MagicAuthProvider + useMagicAuth
 
+- Updated dependencies []:
+  - @magusn/react@0.4.2
+
+## 0.4.1
+
+### Patch Changes
 
 - build(eslint): use @magusn/eslint-config-magusn